refactor: split 4 schema files and jsonb/advanced.ts into focused sub-modules
Easy-win modularity splits:
- schemas/stats.ts (1145 lines) -> stats/{input,output,index}.ts
- schemas/postgis.ts (1097 lines) -> postgis/{basic,advanced,index}.ts
- schemas/jsonb.ts (957 lines) -> jsonb/{basic,advanced,index}.ts
- schemas/core.ts (865 lines) -> core/{queries,transactions,index}.ts
- tools/jsonb/advanced.ts (941 lines) -> {transform,analytics}.ts

Fix 9 stale test imports from monolith -> barrel paths.
