+++ src/adapters/postgresql/schemas/core/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Core schemas barrel - re-exports queries + transactions sub-modules.
+ */
+export * from "./queries.js";
+export * from "./transactions.js";
