+++ src/adapters/postgresql/schemas/jsonb/index.ts
@@ -1,0 +1,5 @@
+/**
+ * JSONB schemas barrel - re-exports basic + advanced sub-modules.
+ */
+export * from "./basic.js";
+export * from "./advanced.js";
