+++ src/adapters/postgresql/schemas/postgis/index.ts
@@ -1,0 +1,5 @@
+/**
+ * PostGIS schemas barrel - re-exports basic + advanced sub-modules.
+ */
+export * from "./basic.js";
+export * from "./advanced.js";
