+++ src/adapters/postgresql/schemas/stats/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Stats schemas barrel - re-exports input + output sub-modules.
+ */
+export * from "./input.js";
+export * from "./output.js";
